+++ automation_tasks_rs/src/secrets_always_local_mod.rs
@@ -1,6 +1,6 @@
 // secrets_always_local_mod.rs
 
-/// Secrets like GitHub API secret_token, crates.io secret_token, SSH private key passphrase and similar
+/// Secrets like GitHub API secret_token, crates.io secret token, docker hub secret_token, SSH private key passphrase and similar
 /// must never go out of this crate. Never pass any secret to an external crate library as much as possible.
 /// The user has the source code under his fingers in this crate. So he knows nobody will mess with this code
 /// once he inspected and reviewed it.
@@ -8,11 +8,27 @@
 /// The simple program flow of functions that need secrets is butchered to avoid secrets leaving this crate.
 /// Now it looks like a mess, but the goal is achieved. The secrets never leave this crate.
 
+// region: Public API constants
+// ANSI colors for Linux terminal
+// https://github.com/shiena/ansicolor/blob/master/README.md
+/// ANSI color
+pub const RED: &str = "\x1b[31m";
+/// ANSI color
+pub const GREEN: &str = "\x1b[32m";
+/// ANSI color
+pub const YELLOW: &str = "\x1b[33m";
+/// ANSI color
+pub const BLUE: &str = "\x1b[34m";
+/// ANSI color
+pub const RESET: &str = "\x1b[0m";
+// endregion: Public API constants
+
+pub use cargo_auto_encrypt_secret_lib::EncryptedString;
+pub use secrecy::ExposeSecret;
+
 pub(crate) mod decrypt_mod {
 
-    use cargo_auto_lib::RED;
-    use cargo_auto_lib::RESET;
-    use secrecy::ExposeSecret;
+    use crate::secrets_always_local_mod::*;
 
     /// The secrets must not leave this crate.
     /// They are never going into an external library crate.
@@ -58,12 +74,7 @@ pub(crate) mod decrypt_mod {
 }
 
 pub(crate) mod encrypt_mod {
-
-    use cargo_auto_lib::RED;
-    use cargo_auto_lib::RESET;
-
-    // bring trait to scope
-    use secrecy::ExposeSecret;
+    use crate::secrets_always_local_mod::*;
 
     /// The secrets must not leave this crate.
     /// They are never going into an external library crate.
@@ -109,7 +120,7 @@ pub(crate) mod secrecy_mod {
     //! But I want to encrypt the content, so I will make a wrapper.
     //! The secrets must always be moved to secrecy types as soon as possible.
 
-    use cargo_auto_encrypt_secret_lib::EncryptedString;
+    use crate::secrets_always_local_mod::*;
 
     pub struct SecretEncryptedString {
         encrypted_string: EncryptedString,
@@ -138,17 +149,7 @@ pub(crate) mod secrecy_mod {
 
 pub(crate) mod ssh_mod {
 
-    #[allow(unused_imports)]
-    use cargo_auto_lib::BLUE;
-    use cargo_auto_lib::GREEN;
-    use cargo_auto_lib::RED;
-    use cargo_auto_lib::RESET;
-    use cargo_auto_lib::YELLOW;
-
     use crate::secrets_always_local_mod::*;
-
-    // bring trait into scope
-    use secrecy::ExposeSecret;
 
     pub struct SshContext {
         signed_passcode_is_a_secret: secrecy::SecretVec<u8>,
@@ -302,15 +303,9 @@ pub(crate) mod github_mod {
     //! Instead of the secret_token, I will pass the struct GitHubClient with the trait SendToGitHubApi.
     //! This way, the secret_token will be encapsulated.
 
+    use crate::secrets_always_local_mod::*;
     use cargo_auto_github_lib as cgl;
-
-    use cargo_auto_lib::BLUE;
-    use cargo_auto_lib::RED;
-    use cargo_auto_lib::RESET;
-
     use reqwest::Client;
-    // bring trait into scope
-    use secrecy::ExposeSecret;
 
     /// Struct GitHubClient contains only private fields
     /// This fields are accessible only to methods in implementation of traits.
@@ -402,6 +397,7 @@ pub(crate) mod github_mod {
         }
 
         /// decrypts the secret_token in memory
+        #[allow(dead_code)]
         pub fn decrypt_secret_token_in_memory(&self) -> secrecy::SecretString {
             self.encrypted_token.expose_decrypted_secret(&self.session_passcode)
         }
@@ -492,11 +488,9 @@ pub(crate) mod crates_io_mod {
     //! Instead of the secret_token, I will pass the struct CratesIoClient with the trait SendToCratesIo.
     //! This way, the secret_token will be encapsulated.
 
+    use crate::secrets_always_local_mod::*;
+    use cargo_auto_lib::ShellCommandLimitedDoubleQuotesSanitizer;
     use cargo_auto_lib::ShellCommandLimitedDoubleQuotesSanitizerTrait;
-    use cargo_auto_lib::BLUE;
-    use cargo_auto_lib::RESET;
-
-    // bring trait into scope
 
     /// Struct CratesIoClient contains only private fields
     /// This fields are accessible only to methods in implementation of traits.
@@ -560,6 +554,16 @@ pub(crate) mod crates_io_mod {
                 crates_io_client
             }
 
+            // check if the plain-text file from `podman login` exists and warn the user because it is a security vulnerability.
+            let file_auth = "~/.cargo/credentials.toml";
+            let file_auth = camino::Utf8Path::new(file_auth);
+            // TODO: check for env variable also?
+            let file_auth_expanded = cargo_auto_encrypt_secret_lib::file_path_home_expand(file_auth);
+            let file_auth_expanded = camino::Utf8Path::new(&file_auth_expanded);
+            if file_auth_expanded.exists() {
+                eprintln!("{RED}Security vulnerability: Found the cargo file with plain-text secret_token: {file_auth_expanded}. It would be better to inspect and remove it. {RESET}")
+            }
+
             let encrypted_string_file_path = camino::Utf8Path::new("~/.ssh/crates_io_secret_token_encrypted.txt");
             let encrypted_string_file_path_expanded = cargo_auto_encrypt_secret_lib::file_path_home_expand(encrypted_string_file_path);
 
@@ -590,6 +594,7 @@ pub(crate) mod crates_io_mod {
         }
 
         /// decrypts the secret_token in memory
+        #[allow(dead_code)]
         pub fn decrypt_secret_token_in_memory(&self) -> secrecy::SecretString {
             self.encrypted_token.expose_decrypted_secret(&self.session_passcode)
         }
@@ -598,9 +603,10 @@ pub(crate) mod crates_io_mod {
         ///
         /// This function encapsulates the secret crates.io secret_token.
         /// The client can be passed to the library. It will not reveal the secret_token.
+        #[allow(dead_code)]
         pub fn publish_to_crates_io(&self) {
             // the secret_token is redacted when print on screen
-            crate::cl::ShellCommandLimitedDoubleQuotesSanitizer::new(r#"cargo publish --token "{secret_token}" "#)
+            ShellCommandLimitedDoubleQuotesSanitizer::new(r#"cargo publish --token "{secret_token}" "#)
                 .unwrap_or_else(|e| panic!("{e}"))
                 .arg_secret("{secret_token}", &self.decrypt_secret_token_in_memory())
                 .unwrap_or_else(|e| panic!("{e}"))
@@ -618,11 +624,9 @@ pub(crate) mod docker_hub_mod {
     //! Instead of the secret_token, I will pass the struct DockerHubClient with the trait SendToDockerHub.
     //! This way, the secret_token will be encapsulated.
 
-    use crate::cl::ShellCommandLimitedDoubleQuotesSanitizerTrait;
-    use crate::cl::BLUE;
-    use crate::cl::RESET;
-
-    // bring trait into scope
+    use crate::secrets_always_local_mod::*;
+    use cargo_auto_lib::ShellCommandLimitedDoubleQuotesSanitizer;
+    use cargo_auto_lib::ShellCommandLimitedDoubleQuotesSanitizerTrait;
 
     /// Struct DockerHubClient contains only private fields
     /// This fields are accessible only to methods in implementation of traits.
@@ -686,6 +690,18 @@ pub(crate) mod docker_hub_mod {
                 docker_hub_client
             }
 
+            // check if the plain-text file from `podman login` exists and warn the user because it is a security vulnerability.
+            let file_auth = "${XDG_RUNTIME_DIR}/containers/auth.json";
+            // TODO: check for env variable also?
+            if let Some(xdg_runtime_dir) = std::env::var_os("XDG_RUNTIME_DIR"){
+                let xdg_runtime_dir=xdg_runtime_dir.to_string_lossy().to_string();
+                let file_auth_expanded = file_auth.replace("${XDG_RUNTIME_DIR}", &xdg_runtime_dir);
+                let file_auth_expanded = camino::Utf8Path::new(&file_auth_expanded);
+                if file_auth_expanded.exists() {
+                    eprintln!("{RED}Security vulnerability: Found the docker hub file with plain-text secret_token: {file_auth_expanded}. It would be better to inspect and remove it. {RESET}")
+                }
+            }
+
             // registry: docker.io -> replace dot into "--""
             // username: bestiadev
             let registry_escaped = registry.replace(".", "--");
@@ -732,7 +748,7 @@ pub(crate) mod docker_hub_mod {
         #[allow(dead_code)]
         pub fn push_to_docker_hub(&self, image_url: &str, user_name: &str) {
             // the secret_token can be used in place of the password in --cred
-            crate::cl::ShellCommandLimitedDoubleQuotesSanitizer::new(r#"podman push --creds "{user_name}:{secret_token}" "{image_url}" "#)
+            ShellCommandLimitedDoubleQuotesSanitizer::new(r#"podman push --creds "{user_name}:{secret_token}" "{image_url}" "#)
                 .unwrap_or_else(|e| panic!("{e}"))
                 .arg("{user_name}", user_name)
                 .unwrap_or_else(|e| panic!("{e}"))